requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.7.0
Flask-CORS==4.0.1
Werkzeug==3.0.3
SQLAlchemy==2.0.31

# AI / ML
transformers==4.44.2
torch==2.4.1
Pillow==10.4.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38

# Voice
gTTS==2.5.3

# Utilities
python-dotenv==1.0.1
marshmallow==3.21.3
email-validator==2.2.0

# Testing
pytest==8.3.2
pytest-flask==1.3.0
